+++ src/components/SalarySettlement/SalaryTable/SalaryTable.tsx
@@ -1,35 +1,9 @@
-import { MAX_SALARY } from '@/constants/salarySettlementPage';
+import { SalaryTableHeader } from '@/components/SalarySettlement/SalaryTable/SalaryTableHeader';
+import { useFilteredEmployees } from '@/hooks/SalarySettlement/useFilteredEmployees';
+import { useSalarySummary } from '@/hooks/SalarySettlement/useSalarySummary';
+import { useEmployees } from '@/hooks/SalarySettlement/useSettlementEmployee';
 import { IEmployeeSalarySettlement } from '@/models/salarySettlement.model';
 import { Checkbox, Table, TextField } from '@radix-ui/themes';
-import { useMemo, useState } from 'react';
-
-// 급여합계 계산로직
-const calculateTotalPay = (employee: IEmployeeSalarySettlement): number => {
-  const {
-    basePay,
-    comprehensiveOvertimePay,
-    annualLeavePay,
-    holidayPay,
-    jobAllowance,
-    incentive,
-    attendanceDeduction,
-    previousMonthUnpaid,
-    overtimePay,
-    weekendWorkPay,
-  } = employee;
-  return (
-    basePay +
-    comprehensiveOvertimePay +
-    annualLeavePay +
-    holidayPay +
-    jobAllowance +
-    incentive -
-    attendanceDeduction +
-    previousMonthUnpaid +
-    overtimePay +
-    weekendWorkPay
-  );
-};
 
 interface SalaryTableProps {
   salarySettlementData: IEmployeeSalarySettlement[];
@@ -44,125 +18,20 @@ export default function SalaryTable({
   selectedPart,
   selectedJob,
 }: SalaryTableProps) {
-  const [employees, setEmployees] = useState(
-    salarySettlementData.map((employee) => ({
-      ...employee,
-      incentive: employee.incentive || 0,
-      previousMonthUnpaid: employee.previousMonthUnpaid || 0,
-      totalPay: calculateTotalPay({
-        ...employee,
-        incentive: employee.incentive || 0,
-        previousMonthUnpaid: employee.previousMonthUnpaid || 0,
-      }),
-      isSelected: false,
-    }))
+  const { employees, handleAllowance, handleCheckbox, handleAllCheckbox } =
+    useEmployees(salarySettlementData);
+  const filteredEmployees = useFilteredEmployees(
+    employees,
+    selectedRegion,
+    selectedPart,
+    selectedJob
   );
-
-  const filteredEmployees = useMemo(() => {
-    return employees.filter(
-      (employee) =>
-        (!selectedRegion || employee.region === selectedRegion) &&
-        (!selectedPart || employee.department === selectedPart) &&
-        (!selectedJob || employee.job === selectedJob)
-    );
-  }, [employees, selectedRegion, selectedPart, selectedJob]);
-
-  const calculateSummary = useMemo(() => {
-    const selectedEmployees = filteredEmployees.filter((emp) => emp.isSelected);
-    const employeesToSum = selectedEmployees.length > 0 ? selectedEmployees : filteredEmployees;
-
-    const totalEmployees = employeesToSum.length;
-    const totalSalary = employeesToSum.reduce((sum, emp) => sum + emp.salary, 0);
-    const totalPay = employeesToSum.reduce((sum, emp) => sum + (emp.totalPay || 0), 0);
-    const totalPreviousMonthUnpaid = employeesToSum.reduce(
-      (sum, emp) => sum + emp.previousMonthUnpaid,
-      0
-    );
-    const totalOvertimePay = employeesToSum.reduce((sum, emp) => sum + emp.overtimePay, 0);
-    const totalWeekendWorkPay = employeesToSum.reduce((sum, emp) => sum + emp.weekendWorkPay, 0);
-    return {
-      totalEmployees,
-      totalSalary,
-      totalPay,
-      totalPreviousMonthUnpaid,
-      totalOvertimePay,
-      totalWeekendWorkPay,
-    };
-  }, [filteredEmployees]);
-
-  const handleInputChange = (
-    index: number,
-    field: 'incentive' | 'previousMonthUnpaid',
-    value: string
-  ) => {
-    const numericValue = value === '' ? 0 : parseInt(value) || 0;
-
-    if (numericValue > MAX_SALARY) {
-      return;
-    }
-
-    const updatedEmployees = employees.map((employee, i) => {
-      if (i === index) {
-        const updatedEmployee = {
-          ...employee,
-          [field]: numericValue,
-        };
-        return {
-          ...updatedEmployee,
-          totalPay: calculateTotalPay(updatedEmployee),
-        };
-      }
-      return employee;
-    });
-    setEmployees(updatedEmployees);
-  };
-
-  const handleCheckbox = (index: number) => {
-    setEmployees(
-      employees.map((emp, i) => (i === index ? { ...emp, isSelected: !emp.isSelected } : emp))
-    );
-  };
-
-  const handleAllCheckbox = () => {
-    const allSelected = employees.every((emp) => emp.isSelected);
-    setEmployees(employees.map((emp) => ({ ...emp, isSelected: !allSelected })));
-  };
+  const calculateSummary = useSalarySummary(filteredEmployees);
 
   return (
     <div className="justify-center items-start min-h-full pb-10">
       <Table.Root className="w-full text-sm text-gray-500">
-        <Table.Header className="text-xs text-gray-700 uppercase bg-gray-10 whitespace-nowrap">
-          <Table.Row align="center">
-            <Table.ColumnHeaderCell>
-              <Checkbox
-                size="1"
-                color="gray"
-                checked={employees.every((emp) => emp.isSelected)}
-                onCheckedChange={handleAllCheckbox}
-              />
-            </Table.ColumnHeaderCell>
-            <Table.ColumnHeaderCell justify="center" className="w-24">
-              이름/
-              <br />
-              근무파트
-            </Table.ColumnHeaderCell>
-            <Table.ColumnHeaderCell justify="center">입사일/퇴사일</Table.ColumnHeaderCell>
-            <Table.ColumnHeaderCell justify="center">월급</Table.ColumnHeaderCell>
-            <Table.ColumnHeaderCell justify="center">기본급</Table.ColumnHeaderCell>
-            <Table.ColumnHeaderCell justify="center">포괄연장수당</Table.ColumnHeaderCell>
-            <Table.ColumnHeaderCell justify="center">연차수당</Table.ColumnHeaderCell>
-            <Table.ColumnHeaderCell justify="center">휴일수당</Table.ColumnHeaderCell>
-            <Table.ColumnHeaderCell justify="center">직무수당</Table.ColumnHeaderCell>
-            <Table.ColumnHeaderCell justify="center">인센티브</Table.ColumnHeaderCell>
-            <Table.ColumnHeaderCell justify="center">근태공제</Table.ColumnHeaderCell>
-            <Table.ColumnHeaderCell justify="center">전월미지급</Table.ColumnHeaderCell>
-            <Table.ColumnHeaderCell justify="center">O.T 수당</Table.ColumnHeaderCell>
-            <Table.ColumnHeaderCell justify="center">주말근로수당</Table.ColumnHeaderCell>
-            <Table.ColumnHeaderCell justify="center" className="w-32">
-              급여합계
-            </Table.ColumnHeaderCell>
-          </Table.Row>
-        </Table.Header>
+        <SalaryTableHeader employees={employees} handleAllCheckbox={handleAllCheckbox} />
 
         <Table.Body>
           {filteredEmployees.map((employee, index) => (
@@ -206,7 +75,7 @@ export default function SalaryTable({
                   className="w-20"
                   size="1"
                   value={employee.incentive === 0 ? '' : employee.incentive.toString()}
-                  onChange={(e) => handleInputChange(index, 'incentive', e.target.value)}
+                  onChange={(e) => handleAllowance(index, 'incentive', e.target.value)}
                   placeholder="0"
                   maxLength={9}
                 />
@@ -223,7 +92,7 @@ export default function SalaryTable({
                       ? ''
                       : employee.previousMonthUnpaid.toString()
                   }
-                  onChange={(e) => handleInputChange(index, 'previousMonthUnpaid', e.target.value)}
+                  onChange={(e) => handleAllowance(index, 'previousMonthUnpaid', e.target.value)}
                   placeholder="0"
                   maxLength={9}
                 />
