+++ src/models/salarySettlement.model.ts
@@ -3,6 +3,7 @@ export interface IEmployeeSalarySettlement {
   name: string;
   region: string;
   department: string;
+  isSelected: boolean;
   job: '의사' | '간호사' | '기타';
   hireDate: string;
   resignDate?: string;
