+++ src/pages/holidayCalendar/HolidayCalendarPage.tsx
@@ -1,4 +1,5 @@
 import { ChangeMonth } from '@/components/Common/ChangeMonth';
+import SelectBox from '@/components/Common/Select';
 import { ChangeWeek } from '@/components/HolidayCalendar/ChangeWeek';
 import DayOffCalendar from '@/components/HolidayCalendar/HolidayCalendar';
 import { Button } from '@radix-ui/themes';
@@ -22,9 +23,15 @@ export default function DayOffCalendarPage() {
   };
 
   return (
-    <div className="flex flex-col gap-4 h-full">
+    <div className="flex flex-col gap-4 h-full w-full">
       <header className="relative flex justify-between items-center">
-        <div>드롭박스</div>
+        <SelectBox
+          title="지점 선택"
+          name="지점 선택"
+          options={DUMMY_DROP_DOWN_MENU}
+          size="small"
+          border={false}
+        />
         {view === 'dayGridMonth' ? (
           <div className="absolute left-1/2 transform -translate-x-1/2">
             <ChangeMonth currentDate={currentDate} onChangeMonth={handleChangeMonth} />
@@ -47,9 +54,26 @@ export default function DayOffCalendarPage() {
         </div>
       </header>
 
-      <div className="flex-grow h-full overflow-y-auto">
+      <div className="flex-grow overflow-y-auto bg-white">
         <DayOffCalendar currentDate={currentDate} view={view} />
       </div>
     </div>
   );
 }
+
+const DUMMY_DROP_DOWN_MENU = [
+  {
+    id: 1,
+    name: '뮤즈의원(강남점)',
+    action: () => {
+      console.log('뮤즈의원(강남점)');
+    },
+  },
+  {
+    id: 2,
+    name: '뮤즈의원(수원인계점)',
+    action: () => {
+      console.log('뮤즈의원(수원인계점)');
+    },
+  },
+];
