+++ src/utils/calculateSettlementTotal.ts
@@ -1,0 +1,28 @@
+import { IEmployeeSalarySettlement } from '@/models/salarySettlement.model';
+
+export const calculateTotalPay = (employee: IEmployeeSalarySettlement): number => {
+  const {
+    basePay,
+    comprehensiveOvertimePay,
+    annualLeavePay,
+    holidayPay,
+    jobAllowance,
+    incentive,
+    attendanceDeduction,
+    previousMonthUnpaid,
+    overtimePay,
+    weekendWorkPay,
+  } = employee;
+  return (
+    basePay +
+    comprehensiveOvertimePay +
+    annualLeavePay +
+    holidayPay +
+    jobAllowance +
+    incentive -
+    attendanceDeduction +
+    previousMonthUnpaid +
+    overtimePay +
+    weekendWorkPay
+  );
+};
